Removing stack name as tag (#326)
Closes https://github.com/chanzuckerberg/cryoet-data-portal/issues/325

### Description

The cluster uses an image if it already exists in the machine. Using the
stack name as the tag causes it to use the old image even if a new image
is available, as there is no difference in their tags. Removing the
explicit static tag makes happy default to the github sha as a tag. This
will fix the caching issue.