feat: Update workflow for automating prod deploy (#471)
Relates to
https://github.com/chanzuckerberg/cryoet-data-portal/issues/438

### Changes introduced

- Adds conventional commit check to PRs as the `release-please`
dependency that we use for generating change logs expects commit
messages to follow the conventional commit format.

#### Prod deploy
- Adds a schedule to deploy on weekdays Mon-Fri @ 18:00 UTC (10AM PST
[winter] / 11AM PDT [summer])
- Conditionals checks introduced before prod deployment:
- If there have been new changes added to the front end since the last
release using change release.
  - If the staging e2e tests pass before kicking off a deploy.
- If the kill switch for prod deploys `secrets.IS_DEPLOY_ALLOWED`, it
allows deployment. The kill switch is added for cases where regressions
are identified in staging that we don't want to be deployed to
production.
- Adds a check to verify if the prod e2e tests pass after kicking off a
deploy


### Doc deploy
- Currently, the doc generation and deploy workflow uses adding a tag as
a trigger. The PR updates it to ignore any tag prefixed by `frontend-`.