docs: Minor updates to documentation site (#611)
Apart from some small typos and such, a meaningful change is removing a
mention of scaling in one of the quick start examples, since tomograms
are not binned any more (noticed during a meeting with Utz).

The updated wording doesn't get into details, but I had two questions:
- Are the tomograms that will get downloaded the highest *available*
resolution, but not necessarily the maximum *possible* resolution?
Answer: Yes.
- When the code iterates over all tomograms, are they unique, or can
there be different resolutions of the same tomogram? Answer: There can
be, but there are none right now.