refactor: using page object model for e2e tests (#856)
This PR tries to use the page object model for the metadataDrawer tests.
Note: some tests are failing - I will need to add Jeremy's ground truth
changes into this change before merging.

Some pros/cons I found along the way:
Pro:
- putting helpers in the page object model forced me to come up with
intelligible names for what was going on

Con:
- the tests are very similar to each other - I'm not sure I properly
balanced being clear with being succinct.

What do we think of this approach?
Is it more readable? Is it harder to track things down?
Do we want less repetition? - if so, we could do this within the page
object model or opt for sticking with functions.