chore: add hygen (#877)
Resolves #875 

Adds code generation for creating new end to end test suites. This
creates skeletons for the following files
```
- newFeature.test.ts
- pageObjects
  - newFeature
    - newFeaturePage.ts
    - types.ts
    - utils.ts
```
Does not add `newFeatureAggregators.ts` for now (I will add this when I
have a better idea of what will live here)

Added instructions for using hygen to the testing wiki page:
https://github.com/chanzuckerberg/cryoet-data-portal/wiki/Testing

It's currently working with the terminal command, but not working with
the vscode extension - after debugging for a few hours, I think this is
good enough for now.