refactor: filter tests page object model (#885)
Related to #832 

This PR refactors the filter tests to use the page object model. The
tests are a bit more involved than the `metadataDrawer`, so this PR also
introduces an `actor` layer that aggregates steps from the `page object`
layer. The goal is to have each test contain a readable set of steps -
enough that a new reader can easily understand what is being tested, but
not so much that it's overwhelming. There should be clear places to look
for what's happening under the hood.