refactor: filter tests (#952)
Relates to #832 

This refactors the rest of the filter tests to use the page object
model. The test files are much more verbose, but also easier to read.
This format plays nicely with VS Code's Playwright extension and devs
can now run one subtest at a time for easier debugging. Co-pilot was
also helpful with the repetitive nature of the tests.

I'm planning to do a casual tech talk on the approach and tooling. LMK
if a walk through would be helpful for reviewing.

<img width="503" alt="Screenshot 2024-07-26 at 12 04 58 PM"
src="https://github.com/user-attachments/assets/8349a360-d36a-4729-b866-2387650d50f9">