feat: Update cell strain links. (#968)
Resolves #912 

Rather than adding additional logic to the InfoLink component that would
only be used for cell strains, this adds a specific CellStrainInfoLink.
This PR also moves the InfoLink components to their own files.

Lastly, this PR uses the fractal file structure - it fits here anyway,
but it would be great to know if there's broader buy in for using this
pattern. If so, I'll move over the Hygen template from CZID so we won't
have to manually set up the folder/file structure for each component we
add.