feat: Generate Python client code using GraphQL introspection (#1008)
Closes #894

This adds a custom code generator for the Python client. The mapping
between GraphQL types, model types, and fields is somewhat explicit, but
doesn't require too much code. The code generation updates the schema
and then parses it using de-factor GraphQL python library (which we
already depend on). We then use Jinja2 templates to populate the models
module.

I [tried some existing packages for generating Python GraphQL
clients](https://github.com/chanzuckerberg/cryoet-data-portal/issues/894#issuecomment-2239561088),
and my sense was that writing something specific for this client is
probably easier than learning how to configure those tools for this
particular client. This approach also has the advantage not imposing its
own client and query style, which makes this compatible with the
existing API by design.

I have done a little manual testing, but testing is mostly through the
newly modified or added Python tests. These include checking that
missing item relationships are handled gracefully by the client, which
is at least a temporary requirement as the ingestion runs and may be a
more permanent one.

As the prod GraphQL server has been updated ahead of the Python client,
these changes also include the corresponding changes to the models
module and GraphQL schema file. As well as the associated test
infrastructure changes.