docs: Move documentation site to Sphinx Immaterial (#1087)
Hi all,

This PR implements the move from the existing Sphinx documentation theme
to [sphinx-immaterial](https://jbms.github.io/sphinx-immaterial/) for
the documentation website.

This is still a work in progress. So far, this PR:
- Closes #944, 
- Closes #889 
- Closes #888 
- Addresses most of #765 

Pending are a few dark mode fixes, some API documentation page
improvements and updates to any graphics on the site.

To see a rendered/live version of this PR, go to
https://melissawm.github.io/cryoet-data-portal

Note: I plan on squashing the commits before merging, let me know what
is the preferred workflow for that. I am also keeping this as a draft
for now because more changes are incoming.

---------

Co-authored-by: Kira Evans <contact@kne42.me>