feat: Support a Client object being shared between threads. (#1351)
This PR adds the ability for a Client object to be shared between
threads. Initially, we created one HTTPTransport object that was used
for the lifetime of a Client. These transports are not thread safe and
can't handle concurrent access!

This PR creates a single Schema object when a Client is instantiated
(this takes ~1/10s so we want to minimize how often we redo this work)
and then creates a new Transport for each request, using the Schema
object we created earlier. Since each request has its own Transport,
it's no longer possible for multiple threads to attempt concurrent
access to a Transport.